fix(playerbot): Continue resolving covariant return types
- Moved RolePerformance, RoleStatistics to IRoleAssignment.h
- Moved DungeonMetrics to IDungeonBehavior.h
- Removed incorrect override marker from LFGGroupCoordinator::TeleportPlayerToDungeon
- All structs returned by value now defined in interface headers
