+++ src/modules/Playerbot/Core/DI/Interfaces/IDungeonBehavior.h
@@ -13,6 +13,8 @@
 #include "Position.h"
 #include <string>
 #include <vector>
+#include <atomic>
+#include <chrono>
 
 // Forward declarations
 class Player;
@@ -25,10 +27,37 @@ namespace Playerbot
 // Forward declarations for nested types
 struct DungeonData;
 struct DungeonEncounter;
-struct DungeonMetrics;
 enum class DungeonRole : uint8;
 enum class EncounterStrategy : uint8;
 enum class ThreatManagement : uint8;
+
+// DungeonMetrics definition (needs full definition for return by value)
+struct DungeonMetrics
+{
+    ::std::atomic<uint32> dungeonsCompleted{0};
+    ::std::atomic<uint32> dungeonsAttempted{0};
+    ::std::atomic<uint32> encountersCompleted{0};
+    ::std::atomic<uint32> encounterWipes{0};
+    ::std::atomic<float> averageCompletionTime{2700000.0f}; // 45 minutes
+    ::std::atomic<float> successRate{0.85f};
+    ::std::atomic<float> encounterSuccessRate{0.9f};
+    ::std::atomic<uint32> totalDamageDealt{0};
+    ::std::atomic<uint32> totalHealingDone{0};
+    ::std::chrono::steady_clock::time_point lastUpdate;
+
+    void Reset() {
+        dungeonsCompleted = 0; dungeonsAttempted = 0; encountersCompleted = 0;
+        encounterWipes = 0; averageCompletionTime = 2700000.0f; successRate = 0.85f;
+        encounterSuccessRate = 0.9f; totalDamageDealt = 0; totalHealingDone = 0;
+        lastUpdate = ::std::chrono::steady_clock::now();
+    }
+
+    float GetCompletionRate() const {
+        uint32 attempted = dungeonsAttempted.load();
+        uint32 completed = dungeonsCompleted.load();
+        return attempted > 0 ? (float)completed / attempted : 0.0f;
+    }
+};
 
 /**
  * @brief Interface for comprehensive dungeon behavior automation
