+++ src/modules/Playerbot/Group/RoleAssignment.h
@@ -169,34 +169,7 @@ public:
     void SetRoleFlexibility(bool isFlexible) override;
     void AddRoleConstraint(GroupRole role, RoleCapability capability) override;
 
-    // Role performance tracking
-    struct RolePerformance
-    {
-        std::atomic<uint32> assignmentsAccepted{0};
-        std::atomic<uint32> assignmentsDeclined{0};
-        std::atomic<float> performanceRating{5.0f};
-        std::atomic<uint32> successfulEncounters{0};
-        std::atomic<uint32> failedEncounters{0};
-        std::atomic<float> averageEffectiveness{0.5f};
-        std::chrono::steady_clock::time_point lastPerformanceUpdate;
-
-        void Reset() {
-            assignmentsAccepted = 0; assignmentsDeclined = 0; performanceRating = 5.0f;
-            successfulEncounters = 0; failedEncounters = 0; averageEffectiveness = 0.5f;
-            lastPerformanceUpdate = std::chrono::steady_clock::now();
-        }
-
-        float GetAcceptanceRate() const {
-            uint32 total = assignmentsAccepted.load() + assignmentsDeclined.load();
-            return total > 0 ? (float)assignmentsAccepted.load() / total : 1.0f;
-        }
-
-        float GetSuccessRate() const {
-            uint32 total = successfulEncounters.load() + failedEncounters.load();
-            return total > 0 ? (float)successfulEncounters.load() / total : 0.5f;
-        }
-    };
-
+    // Role performance tracking (RolePerformance defined in IRoleAssignment.h interface)
     RolePerformance GetPlayerRolePerformance(GroupRole role) override;
     void UpdateRolePerformance(GroupRole role, bool wasSuccessful, float effectiveness) override;
 
@@ -210,30 +183,7 @@ public:
     std::vector<uint32> FindEmergencyReplacements(GroupRole role, uint32 minLevel, uint32 maxLevel) override;
     void HandleRoleEmergency(Group* group, uint32 disconnectedPlayerGuid) override;
 
-    // Role statistics and monitoring
-    struct RoleStatistics
-    {
-        std::atomic<uint32> totalAssignments{0};
-        std::atomic<uint32> successfulAssignments{0};
-        std::atomic<uint32> roleConflicts{0};
-        std::atomic<uint32> emergencyFills{0};
-        std::atomic<float> averageCompositionScore{5.0f};
-        std::atomic<float> roleDistributionEfficiency{0.8f};
-        std::chrono::steady_clock::time_point lastStatsUpdate;
-
-        void Reset() {
-            totalAssignments = 0; successfulAssignments = 0; roleConflicts = 0;
-            emergencyFills = 0; averageCompositionScore = 5.0f; roleDistributionEfficiency = 0.8f;
-            lastStatsUpdate = std::chrono::steady_clock::now();
-        }
-
-        float GetSuccessRate() const {
-            uint32 total = totalAssignments.load();
-            uint32 successful = successfulAssignments.load();
-            return total > 0 ? (float)successful / total : 0.0f;
-        }
-    };
-
+    // Role statistics and monitoring (RoleStatistics defined in IRoleAssignment.h interface)
     RoleStatistics GetGlobalRoleStatistics() override;
     void UpdateRoleStatistics() override;
